add accesstoken to authcontext state

+++ web/src/features/audio/hooks/mutations/useRemoveAudio.ts
@@ -1,10 +1,12 @@
 import { useMutation, useQueryClient } from 'react-query';
+import { useAuth } from '~/contexts/AuthContext';
 import api from '~/utils/api';
 
 
 export function useRemoveAudio(id: number) {
   const queryClient = useQueryClient();
-  const removeAudio = async () => await api.delete(`audios/${id}`);
+  const { accessToken } = useAuth();
+  const removeAudio = async () => await api.delete(`audios/${id}`, { accessToken });
 
   return useMutation(removeAudio, {
     onSuccess() {
